+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changes
 
+## master
+
+- Add FromK8sClient() to get an implementation of the Client interface backed
+  by a *k8s.Client
+
 ## v1.0.0
 
 - Initial release